requirements.txt
pandas
scikit-learn
tensorflow
flask
plotly
psycopg2
kafka-python
hdfs
pyspark
dash
